refactor: extract CieID logic into separate hook
